Changed entities id to int, added files app service

+++ aspnet-core/src/TicketTracker.Core/Entities/Component.cs
@@ -8,13 +8,13 @@ using System.Threading.Tasks;
 using TicketTracker.Authorization.Users;
 
 namespace TicketTracker.Entities {
-    public class Component : FullAuditedEntity<long, User> {
+    public class Component : FullAuditedEntity<int, User> {
         [Required]
         public string Name { get; set; }
         public string Description { get; set; }
 
         public Project Project { get; set; }
-        public long ProjectId { get; set; }
+        public int ProjectId { get; set; }
 
         public List<Ticket> Tickets { get; set; }
     }
